AssetManager: Macro enum identifiers and AssetManifest
This change accomplishes two things:

1) Creates enums from our asset manifest to make it easier to manage our assets.

2) Loads assets from our directory structure based on our AssetManifest.json

These goals help streamline our assets pipeline and make it easy, safe, and optionally dynamic to access and reference our assets.

In the future, as the AssetManager evolves, we will refrain from loading everything at the beginning of runtime and only get what we need for the scene.